requirements.txt
python-dotenv==0.15.0
fastapi
plaid
uvicorn[standard]
SQLAlchemy==2.0.23
fastapi==0.104.1
intuit-oauth
